+++ Battleships/Battleships/Gameboard.h
@@ -26,9 +26,9 @@ public:
 	int getBoardSize()const;
 
 	void initialize();
-	bool addShip(size_t shipSize,  string coord, string dir);
-	bool addShip(size_t shipSize, int x, int y, string dir);
-	bool addShipQuiet(size_t shipSize, int x, int y, string dir);
+	bool addShip(int shipSize,  string coord, string dir);
+	bool addShip(int shipSize, int x, int y, string dir);
+	bool addShipQuiet(int shipSize, int x, int y, string dir);
 	bool shoot(string coord);
 	void sinkShip(int shipIndes);
 	void printShots() const;
